backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-yasg==1.21.7
Pillow==10.4.0
mysqlclient==2.2.4
gunicorn==23.0.0
whitenoise==6.8.2
celery==5.4.0
redis==5.2.0
meilisearch==0.31.0
uvicorn==0.32.1
psutil==6.0.0  # For detailed CPU monitoring and optimization
